distcc-masq-client: Prevent recursion when ccache/distcc invoked directly.
Some broken build processes (e.g. fwbuilder, fossil) try to call ccache and/or distcc explicitly
despite everything in the manpages saying why this is a bad idea.  This results in recursive calls
to ccache/distcc when the compiler is masqueraded.  We can't prevent this when ccache or distcc is
given an explicit /usr/bin pathname, but we can break the cycle if PATH is used and the masquerade
directory is first in the PATH.